Remove unused geometry methods
