--- linesman/geometry/__init__.py
@@ -1,3 +1,0 @@
-from .abstract import Vector, StraightLine
-from .planar import Line
-from .spherical import Geodesic
